requirements.txt
azure-functions
azure-ai-formrecognizer
azure-core
requests
